+++ tests/Integration/Execution/DataLoader/ModelRelationLoaderFetcherTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Integration\Support\DataLoader;
+namespace Tests\Integration\Execution\DataLoader;
 
 use Illuminate\Support\Facades\DB;
 use Nuwave\Lighthouse\Execution\DataLoader\ModelRelationFetcher;
@@ -8,7 +8,7 @@ use Tests\DBTestCase;
 use Tests\Utils\Models\Tag;
 use Tests\Utils\Models\Task;
 
-class ModelRelationLoaderPolymorphicTest extends DBTestCase
+class ModelRelationLoaderFetcherTest extends DBTestCase
 {
     protected function setUp(): void
     {