+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `syncWithoutDetaching` option for BelongsToMany and MorphToMany relationships https://github.com/nuwave/lighthouse/pull/1031
 - Add `injectArgs` option to `@can` directive to pass along client defined
   arguments to the policy check https://github.com/nuwave/lighthouse/pull/1043
+- Allow globally turning off relation batch loading through the
+  config option `batchload_relations` https://github.com/nuwave/lighthouse/pull/1059
+- Add `\Nuwave\Lighthouse\Execution\DataLoader\BatchLoader#loadMany()` function https://github.com/nuwave/lighthouse/pull/973
 - Extend `@rename` directive to work with input fields https://github.com/nuwave/lighthouse/issues/521
 
 ### Changed
@@ -19,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add ability to fetch soft deleted model within `@can` directive to validate permissions
   using `@softDeletes` directive. https://github.com/nuwave/lighthouse/pull/1042
 - Improve the error message for missing field resolvers by offering a solution https://github.com/nuwave/lighthouse/pull/1045
+- Throw `DefinitionException` when missing a type in the type registry https://github.com/nuwave/lighthouse/pull/1066
 
 ## [4.6.0](https://github.com/nuwave/lighthouse/compare/v4.5.3...v4.6.0)